Add PNG and zLib sources to tree Resolve some compilation warnings. Refresh GLEW sources. Resolve gs_bezier crash at shutdown - stop logging at tear-down.

+++ src/glu/libnurbs/internals/sorter.cc
@@ -82,7 +82,7 @@ Sorter::qs1( char *a,  char *l )
     unsigned int n;
 
 start:
-    if((n=l-a) <= (unsigned int)es)
+    if((n=(unsigned int) (l-a)) <= (unsigned int)es)
 	    return;
     n = es * (n / (2*es));
     hp = lp = a+n;
